refactor(xagent): use xbase xWallMs instead of a home-rolled wall clock
xbase/time.h already exposes xWallMs() — CLOCK_REALTIME in ms since
the unix epoch, which is exactly the semantics created_at_ms wants.
The previous commit introduced ai_now_unix_ms_() in turn_private.h
and a duplicate now_unix_ms_() in memory_jsonl.c purely because I
forgot to grep xbase first.

Collapse both onto xWallMs, and while at it switch
xAgentSessionMsg::created_at_ms from long long to uint64_t so its
type matches xWallMs's return. Wall-clock ms is always non-negative
and uint64_t is what every other xbase timebase helper already
returns (xMonoMs, xEventLoopNowMs, sidecar_idle_ms in session.h),
so this aligns the whole codebase on one integer width for "ms".

Test assertions and literals follow: ULL suffixes and uint64_t
locals in memory_test / session_test. Persistence format is
unchanged: memory_jsonl still writes ts as a plain JSON number,
now via %llu with an (unsigned long long) cast — still a portable
integer literal that old parsers read fine (and parse_line_
ignores ts anyway).

Side benefit: memory_jsonl.c no longer needs <time.h> or
<sys/time.h>; the file now relies exclusively on xbase for
platform-clock access.

+++ libs/xagent/session.h
@@ -380,7 +380,7 @@ XDEF_STRUCT(xAgentSessionMsg){
   int         tool_result_is_error;
 
   /* Wall-clock unix-ms when this entry was produced. 0 = unknown. */
-  long long   created_at_ms;
+  uint64_t    created_at_ms;
 };
 
 /**
